test/requirements.txt
aiofiles==0.7.0
ajsonrpc==1.1.0
anyio==3.3.4
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
bottle==0.12.19
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
h11==0.12.0
idna==2.10
ifaddr==0.1.7
iniconfig==1.1.1
marshmallow==3.11.1
nanopb==0.4.5.post1
packaging==20.9
platformio==5.2.2
pluggy==0.13.1
protobuf==3.19.0
py==1.10.0
pyelftools==0.27
pyparsing==2.4.7
pyserial==3.5
pytest==6.2.4
pytest-dotenv==0.5.2
requests==2.25.1
semantic-version==2.8.5
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tabulate==0.8.9
toml==0.10.2
urllib3==1.26.4
uvicorn==0.15.0
wsproto==1.0.0
zeroconf==0.36.9
